requirements.txt
colored==1.4.2
Gooey==1.0.8.1
numpy==1.20.3
Pillow==8.2.0
psutil==5.8.0
pygtrie==2.4.2
python-dotenv==0.17.1
selenium==3.141.0
six==1.16.0
urllib3==1.26.5
wxPython==4.1.1
